+++ package.json
@@ -9,10 +9,10 @@
     "ui5"
   ],
   "scripts": {
-    "build": "npm-run-all --sequential build:base build:compatibility build:localization build:theme-base build:icons build:icons-tnt build:main build:fiori",
+    "build": "npm-run-all --sequential build:base build:ie11 build:localization build:theme-base build:icons build:icons-tnt build:main build:fiori",
     "build:localization": "cd packages/localization && yarn build",
     "build:base": "cd packages/base && yarn build",
-    "build:compatibility": "cd packages/compatibility && yarn build",
+    "build:ie11": "cd packages/ie11 && yarn build",
     "build:theme-base": "cd packages/theme-base && yarn build",
     "build:icons": "cd packages/icons && yarn build",
     "build:icons-tnt": "cd packages/icons-tnt && yarn build",
@@ -20,10 +20,10 @@
     "build:fiori": "cd packages/fiori && yarn build",
     "build:playground": "yarn build:main && yarn build:fiori && cd packages/playground && yarn build",
     "build:playground:master": "yarn build:main && yarn build:fiori && cd packages/playground && yarn build:master",
-    "clean": "npm-run-all --sequential clean:base clean:compatibility clean:localization clean:theme-base clean:icons clean:icons-tnt clean:main clean:fiori",
+    "clean": "npm-run-all --sequential clean:base clean:ie11 clean:localization clean:theme-base clean:icons clean:icons-tnt clean:main clean:fiori",
     "clean:localization": "cd packages/localization && yarn clean",
     "clean:base": "cd packages/base && yarn clean",
-    "clean:compatibility": "cd packages/compatibility && yarn clean",
+    "clean:ie11": "cd packages/ie11 && yarn clean",
     "clean:theme-base": "cd packages/theme-base && yarn clean",
     "clean:icons": "cd packages/icons && yarn clean",
     "clean:icons-tnt": "cd packages/icons-tnt && yarn clean",
@@ -34,15 +34,16 @@
     "scopePrepare:main": "cd packages/main && nps scope.prepare",
     "scopePrepare:fiori": "cd packages/fiori && nps scope.prepare",
     "dev:base": "cd packages/base && nps watch",
+    "dev:ie11": "cd packages/ie11 && nps watch",
     "dev:localization": "cd packages/localization && nps watch",
     "dev:main": "cd packages/main && nps dev",
     "dev:fiori": "cd packages/fiori && nps dev",
     "scopeDev:main": "cd packages/main && nps scope.dev",
     "scopeDev:fiori": "cd packages/fiori && nps scope.dev",
-    "start": "npm-run-all --sequential build:base build:compatibility build:localization build:theme-base build:icons build:icons-tnt prepare:main prepare:fiori start:all",
-    "startWithScope": "npm-run-all --sequential build:base build:compatibility build:localization build:theme-base build:icons build:icons-tnt scopePrepare:main scopePrepare:fiori scopeStart:all",
-    "start:all": "npm-run-all --parallel dev:base dev:localization dev:main dev:fiori",
-    "scopeStart:all": "npm-run-all --parallel dev:base dev:localization scopeDev:main scopeDev:fiori",
+    "start": "npm-run-all --sequential build:base build:ie11 build:localization build:theme-base build:icons build:icons-tnt prepare:main prepare:fiori start:all",
+    "startWithScope": "npm-run-all --sequential build:base build:ie11 build:localization build:theme-base build:icons build:icons-tnt scopePrepare:main scopePrepare:fiori scopeStart:all",
+    "start:all": "npm-run-all --parallel dev:base dev:ie11 dev:localization dev:main dev:fiori",
+    "scopeStart:all": "npm-run-all --parallel dev:base dev:ie11 dev:localization scopeDev:main scopeDev:fiori",
     "start:base": "cd packages/base && yarn start",
     "start:main": "cd packages/main && yarn start",
     "start:fiori": "cd packages/fiori && yarn start",
@@ -73,7 +74,7 @@
     "packages/fiori",
     "packages/icons",
     "packages/icons-tnt",
-    "packages/compatibility",
+    "packages/ie11",
     "packages/tools",
     "packages/playground"
   ]
