--- packages/compatibility/src/features/Edge.js
@@ -1,8 +1,0 @@
-// URLSearchParams
-import "url-search-params-polyfill/index.js";
-
-
-// "pseudo mutation observer" fix for nodeValue
-import "../patchNodeValue.js";
-
-import "../LegacyInterface.js";
