+++ packages/fiori/bundle.es5.js
@@ -1,5 +1,5 @@
 // ES5 bundle targets IE11 only
-import "@ui5/webcomponents-compatibility/dist/features/IE11.js";
+import "@ui5/webcomponents-ie11/dist/features/IE11.js";
 
 import testAssets from "./bundle.esm.js";
