+++ packages/ie11/.eslintignore
