+++ packages/ie11/config/.eslintrc.js
