+++ packages/ie11/src/DOMObserver.js
