+++ packages/ie11/src/LegacyInterface.js
