+++ packages/ie11/src/features/IE11WithWebComponentsPolyfill.js
