+++ packages/ie11/src/patchNodeValue.js
