+++ packages/ie11/src/whenPolyfillLoaded.js
