+++ packages/tools/lib/init-package/resources/bundle.es5.js
@@ -1,5 +1,5 @@
 // ES5 bundle targets IE11 only
-import "@ui5/webcomponents-base/dist/features/browsersupport/IE11.js";
+import "@ui5/webcomponents-ie11/dist/features/IE11.js";
 
 import "./bundle.esm.js";
